pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trik-hash-py"
version = "1.0.0"
description = "Compute and compare cryptographic hashes"
requires-python = ">=3.10"
license = "MIT"

dependencies = [
    "trikhub-sdk>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["trik_hash_py*"]
